Added aura support, made events public, removed player-specific code
Damage function now uses the new DamageInfo class, making it more flexible if more info is needed in the future.
Health class now supports auras. Attacks are ignored if the character isn't vulnerable to that aura. Players should be vulnerable to all, and enemy attacks should use AuraType.EnemyAtk.
The Health script no longer cares if it's attached the player. That's for other scripts to worry about.