Created Destructible Script
Created a Destructible Script to be applied to objects we want the player to destroy. Since attack functionality has not been implemented yet objects are destroyed when the player collides with them at all.

Required actions for functionality:
Attach 2D collider component. Assign player and damgePerHit with values in inspector. Attach Health and BeltCharacter scripts to the object to be destroyed. Must set a Max Health value in health script and Vulnerable Types to Strike