GameManager and post selector scene create player prefabs after selection
instantiates the character prefabs with the same control scheme and input device used by that player in character selector screen.

ex. if player 1 used controller in selector and player 2 used keyboard, the player 1 prefab created in the new scene will be controlled by controller and the player 2 prefab will be controlled by the keyboard

only support selecting apple since that's the only player prefab we have.