Sprite Fixer
Replaced Billboard.cs with SpriteFixer.cs, which now stretches the image instead of rotating it.
This creates a much better view in the scene editor.
Also applied changes to the player prefab, and slightly changed Player.cs so IsGrounded checks any layers the player collides with, instead of just ground.