Remove Max Simultaneous Attackers
The limit on the number of enemies that can attack at once was causing some issues where the count wasn't getting properly decremented if an enemy died while attacking, leading to soft locks where enemies with the pearry aura would never attack and thus were invincible. Instead of fixing the bug, I decided to just remove it as having a bunch of enemies standing around during combat felt odd to begin with.